Calendar Functionality
Got Calendar to function properly. Need to polish some things up, but at this point it is deployable as long as we guide the users to select the right courses in the correct order.

Next step will be to visibly disable the courses that conflict. Then I want to make the card expand and show course description when clicked on.

None of this is needed until final.